// Presentation/Activitygram_Client/src/constants/types/forms.ts
export interface IEventForm {
    title: string;
    startTime: string;
    endTime: string;
    recurrent: string;
    location: string;
    description: string;
    interests: string;
    preconditions: string;
    initiator: string;
    managers: string;
    invited: string;
    images: string;
    QR: string;
    tags: string;
    status: string;
}